requirements.txt
cairocffi==0.8.0
CairoSVG==1.0.22
cffi==1.10.0
cssselect==1.0.1
Django==1.8
django-bootstrap3==8.2.3
django-crispy-forms==1.6.1
django-libs==1.67.12
django-payslip==0.3.2
html5lib==0.999999999
json-rpc==1.10.3
lxml==3.8.0
pycparser==2.17
Pyphen==0.9.4
python-dateutil==2.6.0
pytz==2017.2
simplejson==3.10.0
six==1.10.0
tinycss2==0.5
WeasyPrint==0.38
webencodings==0.5.1